Moving helpers in a separate file
